requirements.txt
pytest~=8.1.1
pytest-xdist~=3.5.0
selenium~=4.22.0
pytest-html~=4.1.1
allure-pytest~=2.13.3
openpyxl~=3.1.4
geopy~=2.4.1
pytest-rerunfailures==13.0